feat: created largeNavbar component
