+++ apps/homepage/src/app/[lang]/_shared/largeNavbar/index.ts
@@ -1,0 +1,1 @@
+export * from './largeNavbar'
