+++ apps/homepage/src/app/[lang]/_shared/largeNavbar/largeNavbar.module.css
@@ -1,0 +1,16 @@
+.main{
+  padding: 10px 20px;
+  display: flex;
+  user-select: none;
+}
+
+.spanText{
+  color: var(--tertiary-token_100);
+}
+
+.buttonMenu{
+  display: flex;
+  align-content: center;
+  margin-left: auto;
+  margin-right: 10px;
+}
